Document how OAS uses semantic versioning (semver).
The specification itself is versioned using semver, so a new top level section
describes specification versioning, what OAS 3.0 means, how patch levels are used,
and how tools use the version.

The `openapi` field in the root object is a semver string that identifies
which OAS the definition file uses and conforms to.

Added a definition of "OpenAPI definition file" to the Definitions section.

Removed the "OpenAPI Version String" section as it is not needed:
info about the `openapi` field is in the table,
 and info about the spec version is in the OpenAPI Version section.